+++ src/juliaupAutoInstall.ts
@@ -23,8 +23,7 @@ export async function installJuliaOrJuliaup(
     }
     // Options for Julia
     const configurePath = 'Configure path'
-    let message =
-        'Julia is not installed. Do you want to install it automatically using juliaup or manually add the path?'
+    let message = 'Julia is not installed. Do you want to install it using juliaup, the official Julia version manager?'
 
     options.push(configurePath)
 
@@ -40,7 +39,7 @@ export async function installJuliaOrJuliaup(
     }
 
     const choice = await vscode.window.showInformationMessage(
-        'Julia Installation',
+        'Automatically install Julia?',
         { modal: true, detail: message + channelSuffix },
         ...options
     )